+++ libs/cardano-ledger-core/src/Cardano/Ledger/Plutus/CostModels.hs
@@ -19,20 +19,7 @@
 {-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
-module Cardano.Ledger.Alonzo.Scripts (
-  Tag (..),
-  BinaryPlutus (..),
-  AlonzoScript (TimelockScript, PlutusScript),
-  Script,
-  txscriptfee,
-  isPlutusScript,
-  pointWiseExUnits,
-  zipSemiExUnits,
-  validScript,
-  transProtocolVersion,
-  eqAlonzoScriptRaw,
-  translateAlonzoScript,
-
+module Cardano.Ledger.Plutus.CostModels (
   -- * Cost Model
   CostModel,
   CostModelError (..),
@@ -47,64 +34,24 @@ module Cardano.Ledger.Alonzo.Scripts (
   costModelParamNames,
   costModelToMap,
   costModelFromMap,
-  ExUnits (ExUnits, exUnitsMem, exUnitsSteps, ..),
-  ExUnits',
-  Prices (..),
   decodeCostModelFailHard,
   CostModels (..),
   PV1.CostModelApplyError (..),
 )
 where
 
-import Cardano.Ledger.Allegra.Scripts (Timelock, eqTimelockRaw, translateTimelock)
-import Cardano.Ledger.Alonzo.Era (AlonzoEra)
-import Cardano.Ledger.BaseTypes (
-  BoundedRational (unboundRational),
-  NonNegativeInterval,
-  ProtVer (..),
-  boundRational,
- )
 import Cardano.Ledger.Binary (
-  Annotator,
   DecCBOR (decCBOR),
   Decoder,
-  DecoderError (..),
   EncCBOR (encCBOR),
   Encoding,
-  ToCBOR (toCBOR),
-  cborError,
   decodeMapByKey,
   encodeFoldableAsDefLenList,
-  getVersion64,
   ifDecoderVersionAtLeast,
  )
-import Cardano.Ledger.Binary.Coders (
-  Decode (Ann, D, From, Invalid, RecD, SumD, Summands),
-  Encode (Rec, Sum, To),
-  Wrapped (..),
-  decode,
-  encode,
-  (!>),
-  (<!),
-  (<*!),
- )
-import Cardano.Ledger.Binary.Plain (serializeAsHexText)
 import Cardano.Ledger.Binary.Version (natVersion)
-import Cardano.Ledger.Coin (Coin (..))
-import Cardano.Ledger.Core
-import Cardano.Ledger.Crypto (Crypto)
-import Cardano.Ledger.MemoBytes (EqRaw (..))
-import Cardano.Ledger.Plutus.Language (
-  BinaryPlutus (..),
-  Language (..),
-  Plutus (..),
-  guardPlutus,
-  mkLanguageEnum,
-  nonNativeLanguages,
- )
-import Cardano.Ledger.SafeHash (SafeToHash (..))
-import Cardano.Ledger.Shelley.Scripts (nativeMultiSigTag)
-import Cardano.Ledger.TreeDiff (Expr (App), ToExpr (..), defaultExprViaShow)
+import Cardano.Ledger.Plutus.Language (Language (..), mkLanguageEnum, nonNativeLanguages)
+import Cardano.Ledger.TreeDiff (Expr (App), ToExpr (..))
 import Control.DeepSeq (NFData (..), deepseq, rwhnf)
 import Control.Monad (forM, when)
 import Control.Monad.Trans.Writer (WriterT (runWriterT))
@@ -112,208 +59,27 @@ import Data.Aeson (
   FromJSON (..),
   Object,
   ToJSON (..),
-  Value (Array, Object, String),
-  object,
+  Value (Array, Object),
   withObject,
-  (.:),
   (.:?),
-  (.=),
  )
-import qualified Data.Aeson as Aeson (Value)
 import Data.Aeson.Key (fromString)
 import Data.Aeson.Types (Parser)
-import Data.DerivingVia (InstantiatedAt (..))
-import Data.Either (isRight)
-import Data.Int (Int64)
 import Data.Map (Map)
 import qualified Data.Map.Strict as Map
-import Data.Measure (BoundedMeasure, Measure)
-import Data.Scientific (fromRationalRepetendLimited)
-import Data.Semigroup (All (..))
 import Data.Text as T (Text)
-import Data.Word (Word64, Word8)
+import Data.Word (Word8)
 import GHC.Generics (Generic)
 import NoThunks.Class (NoThunks (..), allNoThunks)
-import Numeric.Natural (Natural)
-import PlutusCore.Evaluation.Machine.CostModelInterface (CostModelApplyWarn)
-import qualified PlutusLedgerApi.Common as Plutus (showParamName)
+import PlutusLedgerApi.Common (CostModelApplyWarn, showParamName)
 import qualified PlutusLedgerApi.V1 as PV1 (
   CostModelApplyError (..),
   EvaluationContext,
-  MajorProtocolVersion (MajorProtocolVersion),
   ParamName,
-  ScriptDecodeError,
-  ScriptForEvaluation,
-  deserialiseScript,
   mkEvaluationContext,
  )
-import qualified PlutusLedgerApi.V2 as PV2 (ParamName, deserialiseScript, mkEvaluationContext)
-import qualified PlutusLedgerApi.V3 as PV3 (ParamName, deserialiseScript, mkEvaluationContext)
-
--- | Marker indicating the part of a transaction for which this script is acting
--- as a validator.
-data Tag
-  = -- | Validates spending a script-locked UTxO
-    Spend
-  | -- | Validates minting new tokens
-    Mint
-  | -- | Validates certificate transactions
-    Cert
-  | -- | Validates withdrawal from a reward account
-    Rewrd
-  deriving (Eq, Generic, Ord, Show, Enum, Bounded)
-
-instance NoThunks Tag
-
-instance NFData Tag where
-  rnf = rwhnf
-
-instance ToExpr Tag
-
--- =======================================================
-
--- | Scripts in the Alonzo Era, Either a Timelock script or a Plutus script.
-data AlonzoScript era
-  = TimelockScript !(Timelock era)
-  | PlutusScript !Plutus
-  deriving (Eq, Generic, NoThunks)
-
-instance ToExpr (AlonzoScript era)
-
-translateAlonzoScript ::
-  (Era era1, Era era2, EraCrypto era1 ~ EraCrypto era2) =>
-  AlonzoScript era1 ->
-  AlonzoScript era2
-translateAlonzoScript = \case
-  TimelockScript ts -> TimelockScript $ translateTimelock ts
-  PlutusScript ps -> PlutusScript ps
-
-instance (EraScript era, Script era ~ AlonzoScript era) => Show (AlonzoScript era) where
-  show (TimelockScript x) = "TimelockScript " ++ show x
-  show s@(PlutusScript plutus) =
-    "PlutusScript " ++ show (plutusLanguage plutus) ++ " " ++ show (hashScript @era s)
-
-instance NFData (AlonzoScript era)
-
--- | Both constructors know their original bytes
-instance SafeToHash (AlonzoScript era) where
-  originalBytes (TimelockScript t) = originalBytes t
-  originalBytes (PlutusScript (Plutus _ binaryPlutus)) = originalBytes binaryPlutus
-
-isPlutusScript :: EraScript era => Script era -> Bool
-isPlutusScript = not . isNativeScript
-
-instance Crypto c => EraScript (AlonzoEra c) where
-  type Script (AlonzoEra c) = AlonzoScript (AlonzoEra c)
-  type NativeScript (AlonzoEra c) = Timelock (AlonzoEra c)
-
-  upgradeScript = TimelockScript . translateTimelock
-
-  scriptPrefixTag script =
-    case script of
-      TimelockScript _ -> nativeMultiSigTag -- "\x00"
-      PlutusScript (Plutus PlutusV1 _) -> "\x01"
-      PlutusScript (Plutus PlutusV2 _) -> "\x02"
-      PlutusScript (Plutus PlutusV3 _) -> "\x03"
-
-  getNativeScript = \case
-    TimelockScript ts -> Just ts
-    _ -> Nothing
-
-instance EqRaw (AlonzoScript era) where
-  eqRaw = eqAlonzoScriptRaw
-
-instance Era era => ToJSON (AlonzoScript era) where
-  toJSON = String . serializeAsHexText
-
--- ===========================================
-
--- | Arbitrary execution unit in which we measure the cost of scripts in terms
--- of space in memory and execution time.
---
--- The ledger itself uses 'ExUnits' Natural' exclusively, but the flexibility here
--- allows the consensus layer to translate the execution units into something
--- equivalent to 'ExUnits (Inf Natural)'. This is needed in order to provide
--- a 'BoundedMeasure' instance, which itself is needed for the alonzo instance of
--- 'TxLimits' (in consensus).
-data ExUnits' a = ExUnits'
-  { exUnitsMem' :: !a
-  , exUnitsSteps' :: !a
-  }
-  deriving (Eq, Generic, Show, Functor)
-  -- It is deliberate that there is no Ord instance, use `pointWiseExUnits` instead.
-  deriving
-    (Measure, BoundedMeasure)
-    via (InstantiatedAt Generic (ExUnits' a))
-  deriving
-    (Monoid, Semigroup)
-    via (InstantiatedAt Measure (ExUnits' a))
-
-instance NoThunks a => NoThunks (ExUnits' a)
-
-instance NFData a => NFData (ExUnits' a)
-
-deriving instance ToJSON a => ToJSON (ExUnits' a)
-
-deriving instance FromJSON a => FromJSON (ExUnits' a)
-
--- | This newtype wrapper of ExUnits' is used to hide
---  an implementation detail inside the ExUnits pattern.
-newtype ExUnits = WrapExUnits {unWrapExUnits :: ExUnits' Natural}
-  deriving (Eq, Generic, Show)
-  deriving newtype (Monoid, Semigroup)
-
-instance NoThunks ExUnits
-
-instance NFData ExUnits
-
-instance ToJSON ExUnits where
-  toJSON exUnits@(ExUnits _ _) =
-    let ExUnits {exUnitsMem, exUnitsSteps} = exUnits
-     in object
-          [ "exUnitsMem" .= toJSON exUnitsMem
-          , "exUnitsSteps" .= toJSON exUnitsSteps
-          ]
-
-instance FromJSON ExUnits where
-  parseJSON = withObject "exUnits" $ \o -> do
-    exUnitsMem <- checkWord64Bounds =<< o .: "exUnitsMem"
-    exUnitsSteps <- checkWord64Bounds =<< o .: "exUnitsSteps"
-    pure $ ExUnits {exUnitsMem, exUnitsSteps}
-    where
-      checkWord64Bounds n =
-        if n >= fromIntegral (minBound @Word64)
-          && n <= fromIntegral (maxBound @Word64)
-          then pure n
-          else fail ("Unit out of bounds for Word64: " <> show n)
-
--- | Arbitrary execution unit in which we measure the cost of scripts in terms
--- of space in memory and execution time.
---
--- This pattern hides the fact that ExUnits' is parametric in the underlying type.
--- The ledger itself uses 'ExUnits' Natural' exclusively.
---
--- We would have preferred to use a type alias for 'ExUnits' Natural',
--- but this is not possible: https://gitlab.haskell.org/ghc/ghc/-/issues/19507.
-pattern ExUnits :: Natural -> Natural -> ExUnits
-pattern ExUnits {exUnitsMem, exUnitsSteps} <-
-  WrapExUnits (ExUnits' exUnitsMem exUnitsSteps)
-  where
-    ExUnits m s = WrapExUnits (ExUnits' m s)
-
-{-# COMPLETE ExUnits #-}
-
--- | It is deliberate that there is no `Ord` instance for `ExUnits`. Use this function to
---   compare if one `ExUnit` is pointwise compareable to another. In case when `Ord`
---   instance like comparison is necessary you can use @`zipSemiExUnits` `compare`@
-pointWiseExUnits :: (Natural -> Natural -> Bool) -> ExUnits -> ExUnits -> Bool
-pointWiseExUnits f ex1 ex2 = getAll (zipSemiExUnits (\x y -> All (f x y)) ex1 ex2)
-
--- | Pointwise combine units into a semigroup and mappened the results.
-zipSemiExUnits :: Semigroup a => (Natural -> Natural -> a) -> ExUnits -> ExUnits -> a
-zipSemiExUnits f (ExUnits m1 s1) (ExUnits m2 s2) = (m1 `f` m2) <> (s1 `f` s2)
-
--- =====================================
+import qualified PlutusLedgerApi.V2 as PV2 (ParamName, mkEvaluationContext)
+import qualified PlutusLedgerApi.V3 as PV3 (ParamName, mkEvaluationContext)
 
 -- | A language dependent cost model for the Plutus evaluator.
 -- Note that the `PV1.EvaluationContext` is entirely dependent on the
@@ -407,9 +173,9 @@ plutusV1ParamNames =
         ]
 
 plutusVXParamNames :: Language -> [Text]
-plutusVXParamNames PlutusV1 = Plutus.showParamName <$> [minBound .. maxBound :: PV1.ParamName]
-plutusVXParamNames PlutusV2 = Plutus.showParamName <$> [minBound .. maxBound :: PV2.ParamName]
-plutusVXParamNames PlutusV3 = Plutus.showParamName <$> [minBound .. maxBound :: PV3.ParamName]
+plutusVXParamNames PlutusV1 = showParamName <$> [minBound .. maxBound :: PV1.ParamName]
+plutusVXParamNames PlutusV2 = showParamName <$> [minBound .. maxBound :: PV2.ParamName]
+plutusVXParamNames PlutusV3 = showParamName <$> [minBound .. maxBound :: PV3.ParamName]
 
 validateCostModel :: MonadFail m => Language -> [Integer] -> m CostModel
 validateCostModel lang cmps =
@@ -618,171 +384,8 @@ instance ToJSON CostModels where
 encodeCostModel :: CostModel -> Encoding
 encodeCostModel = encodeFoldableAsDefLenList encCBOR . getCostModelParams
 
--- ==================================
-
--- | Prices per execution unit
-data Prices = Prices
-  { prMem :: !NonNegativeInterval
-  , prSteps :: !NonNegativeInterval
-  }
-  deriving (Eq, Generic, Show, Ord)
-
-instance NoThunks Prices
-
-instance NFData Prices
-
-instance ToJSON Prices where
-  toJSON Prices {prSteps, prMem} =
-    -- We cannot round-trip via NonNegativeInterval, so we go via Rational
-    object
-      [ "prSteps" .= toRationalJSON (unboundRational prSteps)
-      , "prMem" .= toRationalJSON (unboundRational prMem)
-      ]
-
-toRationalJSON :: Rational -> Aeson.Value
-toRationalJSON r =
-  case fromRationalRepetendLimited 20 r of
-    Right (s, Nothing) -> toJSON s
-    _ -> toJSON r
-
-instance FromJSON Prices where
-  parseJSON =
-    withObject "prices" $ \o -> do
-      steps <- o .: "prSteps"
-      mem <- o .: "prMem"
-      prSteps <- checkBoundedRational steps
-      prMem <- checkBoundedRational mem
-      return Prices {prSteps, prMem}
-    where
-      -- We cannot round-trip via NonNegativeInterval, so we go via Rational
-      checkBoundedRational r =
-        case boundRational r of
-          Nothing -> fail ("too much precision for bounded rational: " ++ show r)
-          Just s -> return s
-
--- | Compute the cost of a script based upon prices and the number of execution
--- units.
-txscriptfee :: Prices -> ExUnits -> Coin
-txscriptfee Prices {prMem, prSteps} ExUnits {exUnitsMem = m, exUnitsSteps = s} =
-  Coin $
-    ceiling $
-      (fromIntegral m * unboundRational prMem)
-        + (fromIntegral s * unboundRational prSteps)
-
---------------------------------------------------------------------------------
--- Serialisation
---------------------------------------------------------------------------------
-
-tagToWord8 :: Tag -> Word8
-tagToWord8 = toEnum . fromEnum
-
-word8ToTag :: Word8 -> Maybe Tag
-word8ToTag e
-  | fromEnum e > fromEnum (Prelude.maxBound :: Tag) = Nothing
-  | fromEnum e < fromEnum (minBound :: Tag) = Nothing
-  | otherwise = Just $ toEnum (fromEnum e)
-
-instance EncCBOR Tag where
-  encCBOR = encCBOR . tagToWord8
-
-instance DecCBOR Tag where
-  decCBOR =
-    word8ToTag <$> decCBOR >>= \case
-      Nothing -> cborError $ DecoderErrorCustom "Tag" "Unknown redeemer tag"
-      Just n -> pure n
-  {-# INLINE decCBOR #-}
-
-instance EncCBOR ExUnits where
-  encCBOR (ExUnits m s) = encode $ Rec ExUnits !> To m !> To s
-
-instance DecCBOR ExUnits where
-  decCBOR = decode $ RecD ExUnits <! D decNat <! D decNat
-    where
-      decNat :: Decoder s Natural
-      decNat = do
-        x <- decCBOR
-        when
-          (x > fromIntegral (Prelude.maxBound :: Int64))
-          ( cborError $
-              DecoderErrorCustom "ExUnits field" "values must not exceed maxBound :: Int64"
-          )
-        pure $ wordToNatural x
-      {-# INLINE decNat #-}
-      wordToNatural :: Word64 -> Natural
-      wordToNatural = fromIntegral
-      {-# INLINE wordToNatural #-}
-  {-# INLINE decCBOR #-}
-
-instance EncCBOR Prices where
-  encCBOR (Prices m s) = encode $ Rec Prices !> To m !> To s
-
-instance DecCBOR Prices where
-  decCBOR = decode $ RecD Prices <! From <! From
-  {-# INLINE decCBOR #-}
-
-instance Era era => EncCBOR (AlonzoScript era)
-
-instance Era era => ToCBOR (AlonzoScript era) where
-  toCBOR = toEraCBOR @era . encode . encodeScript
-
-encodeScript :: Era era => AlonzoScript era -> Encode 'Open (AlonzoScript era)
-encodeScript = \case
-  TimelockScript i -> Sum TimelockScript 0 !> To i
-  PlutusScript (Plutus PlutusV1 s) -> Sum (PlutusScript . Plutus PlutusV1) 1 !> To s
-  PlutusScript (Plutus PlutusV2 s) -> Sum (PlutusScript . Plutus PlutusV2) 2 !> To s
-  PlutusScript (Plutus PlutusV3 s) -> Sum (PlutusScript . Plutus PlutusV3) 3 !> To s
-
-instance Era era => DecCBOR (Annotator (AlonzoScript era)) where
-  decCBOR = decode (Summands "Alonzo Script" decodeScript)
-    where
-      decodeAnnPlutus lang =
-        Ann (SumD $ PlutusScript . Plutus lang) <*! Ann (D (guardPlutus lang >> decCBOR))
-      {-# INLINE decodeAnnPlutus #-}
-      decodeScript :: Word -> Decode 'Open (Annotator (AlonzoScript era))
-      decodeScript = \case
-        0 -> Ann (SumD TimelockScript) <*! From
-        1 -> decodeAnnPlutus PlutusV1
-        2 -> decodeAnnPlutus PlutusV2
-        3 -> decodeAnnPlutus PlutusV3
-        n -> Invalid n
-      {-# INLINE decodeScript #-}
-  {-# INLINE decCBOR #-}
-
--- | Test that every Alonzo script represents a real Script.
---     Run deepseq to see that there are no infinite computations and that
---     every Plutus Script unflattens into a real PV1.Script
-validScript :: ProtVer -> AlonzoScript era -> Bool
-validScript pv script =
-  case script of
-    TimelockScript sc -> deepseq sc True
-    PlutusScript (Plutus lang (BinaryPlutus bytes)) ->
-      let deserialiseScript =
-            case lang of
-              PlutusV1 -> PV1.deserialiseScript
-              PlutusV2 -> PV2.deserialiseScript
-              PlutusV3 -> PV3.deserialiseScript
-          eWellFormed :: Either PV1.ScriptDecodeError PV1.ScriptForEvaluation
-          eWellFormed = deserialiseScript (transProtocolVersion pv) bytes
-       in isRight eWellFormed
-
-transProtocolVersion :: ProtVer -> PV1.MajorProtocolVersion
-transProtocolVersion (ProtVer major _minor) =
-  PV1.MajorProtocolVersion ((fromIntegral :: Word64 -> Int) (getVersion64 major))
-
--- | Check the equality of two underlying types, while ignoring their binary
--- representation, which `Eq` instance normally does. This is used for testing.
-eqAlonzoScriptRaw :: AlonzoScript era -> AlonzoScript era -> Bool
-eqAlonzoScriptRaw (TimelockScript t1) (TimelockScript t2) = eqTimelockRaw t1 t2
-eqAlonzoScriptRaw (PlutusScript ps1) (PlutusScript ps2) = ps1 == ps2
-eqAlonzoScriptRaw _ _ = False
-
 instance ToExpr CostModel where
   toExpr (CostModel lang cmmap _) =
     App "CostModel" [toExpr lang, toExpr cmmap, App "PV1.EvaluationContext" []]
 
 instance ToExpr CostModels
-
-instance ToExpr Prices
-
-instance ToExpr ExUnits where
-  toExpr (WrapExUnits (ExUnits' x y)) = App "ExUnits" [defaultExprViaShow x, defaultExprViaShow y]